added AddContact function
added AddContact function.

front-end and back-end teammates were notified of expected variable names. For instance, front-end "input ids" need to be 
firstNameText
lastNameText
phoneText
emailText

and the variables being sent to php are the following
jsNewFirst
jsNewLast
jsNewPhone
jsNewEmail
 
I also reformatted the js to kind of better section off with section headers to help navigate the document. 

I also added clear functions... so stuff in these are commented out and will need to be uncommented as we go. but there purpose is to clear results from other boxes. So if someone does a search prior to adding or deleting or editing when they do the other function the search result field will be cleared.