Back end edits (#13)
* added registration functionality and cleaned up searchContact()

This is a functioning code.js. doLogin, goRegister, doRegister, doLogout, searchContact are all functional. 

we need editContact, deleteContact, and addContact funtionality

* added delete contact semi-functionality

I have got up to searching for the contact and narrowing down... no prompting messages to narrow down... may need to be implemented.

* Can search. Can not delete. Needs to delete also

header says it all

* Update DeleteContact.php

Finalized the delete functionality. js needs to send contactId and userId and to handle the error returns

* Added ID to the return payload

I needed the contact IDs returned to JS so i could use them for delete.

* Fully operational Delete function

delete button is displayed per contact shown.

* added AddContact function

added AddContact function.

front-end and back-end teammates were notified of expected variable names. For instance, front-end "input ids" need to be 
firstNameText
lastNameText
phoneText
emailText

and the variables being sent to php are the following
jsNewFirst
jsNewLast
jsNewPhone
jsNewEmail
 
I also reformatted the js to kind of better section off with section headers to help navigate the document. 

I also added clear functions... so stuff in these are commented out and will need to be uncommented as we go. but there purpose is to clear results from other boxes. So if someone does a search prior to adding or deleting or editing when they do the other function the search result field will be cleared.

---------

Co-authored-by: Daylen Hendricks <131936800+DaylenHendricks@users.noreply.github.com>